Enables relative templatesPath configuration
We are using this extension in our team and the use case is basically to
be able to have some templates committed in our CVS system (so that any
newjoiner can quickly bootstrap a set of templated files).

The problem we're facing is that using a relative path as the
`templatesPath` resolves to an absolute path from which the first vscode
instance was launched (which might not always be the project you're
currently looking at).

This commit makes it so that:

* Given a relative `templatesPath`
* Given an existing `workspaceRoot` directory

It composes the `templatesPath` directory as `path.join(workspaceRoot,
templatesPath)`.